requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
folium==0.15.1
scikit-learn==1.3.2
jupyter==1.0.0
pywebview==4.4.1 